requirements.txt
picamera
gpiozero
google-cloud-storage==1.0.0